Workers now handle integer messages
Workers now receive an integer in the unidirectional message.  If we get
a non-zero, the worker halves it and sends it back.  Workers will loop
until receiving a zero.

On the controller side, we just send everyone a zero.  So, the workers
aren't actually looping right now.  That will change later!