Finally, our README!
This provides a quick-start for Ubuntu users (on Bionic or later), to
quickly see the code run.  Then, I lightly dive into the different
components in use (from MPI down to mpi4py).  I then explain how to run
the code with `mpirun` on a single machine, and with `mpirun` on two
machines.